orders-service/requirements.txt
Flask
pymongo
PyJWT
flask-cors
requests

